+++ lib/src/controls/better_player_controls_state.dart
@@ -3,10 +3,10 @@ import 'dart:math';
 
 // Project imports:
 import 'package:better_player/better_player.dart';
+import 'package:better_player/src/asms/better_player_asms_audio_track.dart';
+import 'package:better_player/src/asms/better_player_asms_track.dart';
 import 'package:better_player/src/controls/better_player_clickable_widget.dart';
 import 'package:better_player/src/core/better_player_utils.dart';
-import 'package:better_player/src/hls/better_player_hls_audio_track.dart';
-import 'package:better_player/src/hls/better_player_hls_track.dart';
 import 'package:better_player/src/video_player/video_player.dart';
 
 // Flutter imports:
@@ -248,26 +248,29 @@ abstract class BetterPlayerControlsState<T extends StatefulWidget>
   }
 
   ///Build both track and resolution selection
-  ///Track selection is used for HLS videos
+  ///Track selection is used for HLS / DASH videos
   ///Resolution selection is used for normal videos
   void _showQualitiesSelectionWidget() {
-    final List<String> trackNames =
-        betterPlayerController!.betterPlayerDataSource!.hlsTrackNames ?? [];
-    final List<BetterPlayerHlsTrack> tracks =
-        betterPlayerController!.betterPlayerTracks;
+    // HLS / DASH
+    final List<String> asmsTrackNames =
+        betterPlayerController!.betterPlayerDataSource!.asmsTrackNames ?? [];
+    final List<BetterPlayerAsmsTrack> asmsTracks =
+        betterPlayerController!.betterPlayerAsmsTracks;
     final List<Widget> children = [];
-    for (var index = 0; index < tracks.length; index++) {
-      final track = tracks[index];
+    for (var index = 0; index < asmsTracks.length; index++) {
+      final track = asmsTracks[index];
 
       String? preferredName;
       if (track.height == 0 && track.width == 0 && track.bitrate == 0) {
         preferredName = betterPlayerController!.translations.qualityAuto;
       } else {
-        preferredName = trackNames.length > index ? trackNames[index] : null;
+        preferredName =
+            asmsTrackNames.length > index ? asmsTrackNames[index] : null;
       }
-      children.add(_buildTrackRow(tracks[index], preferredName));
+      children.add(_buildTrackRow(asmsTracks[index], preferredName));
     }
 
+    // normal videos
     final resolutions =
         betterPlayerController!.betterPlayerDataSource!.resolutions;
     resolutions?.forEach((key, value) {
@@ -276,7 +279,7 @@ abstract class BetterPlayerControlsState<T extends StatefulWidget>
 
     if (children.isEmpty) {
       children.add(
-        _buildTrackRow(BetterPlayerHlsTrack.defaultTrack(),
+        _buildTrackRow(BetterPlayerAsmsTrack.defaultTrack(),
             betterPlayerController!.translations.qualityAuto),
       );
     }
@@ -284,11 +287,16 @@ abstract class BetterPlayerControlsState<T extends StatefulWidget>
     _showModalBottomSheet(children);
   }
 
-  Widget _buildTrackRow(BetterPlayerHlsTrack track, String? preferredName) {
+  Widget _buildTrackRow(BetterPlayerAsmsTrack track, String? preferredName) {
+    final int width = track.width ?? 0;
+    final int height = track.height ?? 0;
+    final int bitrate = track.bitrate ?? 0;
+    final String mimeType = (track.mimeType ?? '').replaceAll('video/', '');
     final String trackName = preferredName ??
-        "${track.width}x${track.height} ${BetterPlayerUtils.formatBitrate(track.bitrate!)}";
+        "${width}x$height ${BetterPlayerUtils.formatBitrate(bitrate)} $mimeType";
 
-    final selectedTrack = betterPlayerController!.betterPlayerTrack;
+    final BetterPlayerAsmsTrack? selectedTrack =
+        betterPlayerController!.betterPlayerAsmsTrack;
     final bool isSelected = selectedTrack != null && selectedTrack == track;
 
     return BetterPlayerMaterialClickableWidget(
@@ -335,23 +343,24 @@ abstract class BetterPlayerControlsState<T extends StatefulWidget>
   }
 
   void _showAudioTracksSelectionWidget() {
-    final List<BetterPlayerHlsAudioTrack>? tracks =
-        betterPlayerController!.betterPlayerAudioTracks;
+    //HLS / DASH
+    final List<BetterPlayerAsmsAudioTrack>? asmsTracks =
+        betterPlayerController!.betterPlayerAsmsAudioTracks;
     final List<Widget> children = [];
-    final BetterPlayerHlsAudioTrack? selectedAudioTrack =
-        betterPlayerController!.betterPlayerAudioTrack;
-    if (tracks != null) {
-      for (var index = 0; index < tracks.length; index++) {
-        final bool isSelected =
-            selectedAudioTrack != null && selectedAudioTrack == tracks[index];
-        children.add(_buildAudioTrackRow(tracks[index], isSelected));
+    final BetterPlayerAsmsAudioTrack? selectedAsmsAudioTrack =
+        betterPlayerController!.betterPlayerAsmsAudioTrack;
+    if (asmsTracks != null) {
+      for (var index = 0; index < asmsTracks.length; index++) {
+        final bool isSelected = selectedAsmsAudioTrack != null &&
+            selectedAsmsAudioTrack == asmsTracks[index];
+        children.add(_buildAudioTrackRow(asmsTracks[index], isSelected));
       }
     }
 
     if (children.isEmpty) {
       children.add(
         _buildAudioTrackRow(
-          BetterPlayerHlsAudioTrack(
+          BetterPlayerAsmsAudioTrack(
             label: betterPlayerController!.translations.generalDefault,
           ),
           true,
@@ -363,7 +372,7 @@ abstract class BetterPlayerControlsState<T extends StatefulWidget>
   }
 
   Widget _buildAudioTrackRow(
-      BetterPlayerHlsAudioTrack audioTrack, bool isSelected) {
+      BetterPlayerAsmsAudioTrack audioTrack, bool isSelected) {
     return BetterPlayerMaterialClickableWidget(
       onTap: () {
         Navigator.of(context).pop();